requirements.txt
six
pyhandle
click
sqlalchemy
mock
bson
pytz
pymongo
ckanext
locust